requirements.txt
pylint 
streamlit
matplotlib
python-docx
wordcloud
PyMuPDF
